--- a/src/utils.jl
+++ b/src/utils.jl
@@ -38,6 +38,16 @@ function atanpos(y, x)
 end
 
 """
+    checklat(lat)
+Checks if the latitude is in the range `[-90°,90°]` and returns it if yes or throws an error otherwise.
+"""
+function checklat(lat)
+  if abs(lat) > 90u"°"
+    throw(ArgumentError("the latitude must be in the range [-90°,90°]"))
+  end
+  lat
+end
+"""
     fixlon(lon)
 
 Fix the longitude to be in the range `[-180°,180°]`.