--- a/src/utils.jl
+++ b/src/utils.jl
@@ -58,7 +58,7 @@ Fix the longitude to be in the range `(-180°,180°]`.
 """
 function fixlon(lon) 
   lon = rem(lon, 360°, RoundNearest)
-  ifelse(lon == -180°, 180°, lon)
+  ifelse(lon == -180°, oftype(lon, 180°), lon)
 end
 
 """