--- a/src/utils.jl
+++ b/src/utils.jl
@@ -45,11 +45,10 @@ The check has a tolerance of `1e-6°` around +90° and -90°, so if `90° < abs(
 """
 function checklat(lat)
   abslat = abs(lat)
-  T = typeof(abslat)
   if abslat > 90° + 1e-6°
     throw(ArgumentError("the latitude must be in the range [-90°,90°], while $lat was provided"))
   end
-  ifelse(abslat > 90°, sign(lat) * T(90), lat)
+  ifelse(abslat > 90°, oftype(lat, sign(lat) * 90°), lat)
 end
 """
     fixlon(lon)