--- a/src/utils.jl
+++ b/src/utils.jl
@@ -46,16 +46,14 @@ end
 
 """
     checklat(lat)
-Checks if the latitude is in the range `[-90°,90°]` and returns it if yes or throws an error otherwise.
-
-The check has a tolerance of `1e-6°` around +90° and -90°, so if `90° < abs(lat) < 90.000001°` the function returns either +90° or -90° depending on the sign of the input.
+Checks if the latitude is in the range `[-90°,90°]`.
 """
 function checklat(lat)
   abslat = abs(lat)
-  if abslat > 90° + 1e-6°
+  if abslat > 90°
     throw(ArgumentError("the latitude must be in the range [-90°,90°], while $lat was provided"))
   end
-  ifelse(abslat > 90°, oftype(lat, sign(lat) * 90°), lat)
+  lat
 end
 
 """