This PR reintroduces the `checklat` function initially proposed in #121 and improves the `fixlon` function to returns a value in the (-180, 180] range rather than the [-180, 180] one.

This removes the need of the `islon180` function as -180 is converted to 180° at construction.

~~The check is now also enforced in the inner constructor of the various geographic crs types to avoid the possibility of inadvertently skip the check.~~
Given that this check indeed introduces overhead, I moved back the check on the more specialized outer constructor, leaving the fully parametrized inner constructor without checks in case one knows already the inputs he is providing are within bounds (like inside custom `rand` methods).

I had to add some tolerance for the out of bounds of checklat (1e-6° at the moment) as unfortunately two tests in back and forth conversion between geographic and projected spaces were breaking due to rounding errors in computation that ledd to latitude values lower but very close to -90° (e.g. `-90.00000000000003°`, see [this CI run log](https://github.com/JuliaEarth/CoordRefSystems.jl/pull/146/checks?check_run_id=28300894477)).

I still believe that having a check on the inputs is important for ensuring consistency, but I would like to know other opinions on how to handle these potential errors caused by rounding problems.

# Benchmark
The performance actually seems to improve in 1.11, and 1.9 (on windows) in my tests with the code below, which does not rely on custom `rand` anymore to avoid bias and implement basically the same methodology for creating random points without the problems in #147:
```julia
using CoordRefSystems
using CoordRefSystems: Deg, Met
using Unitful: °, m
using BenchmarkTools

function randpoint(::Type{T}) where T <: Union{LatLon, AuthalicLatLon, GeocentricLatLon}
    lat = -90° + rand() * 180°
    lon = -360° + rand() * 2 * 360°
    T(lat, lon)
end
function randpoint(::Type{T}) where T <: LatLonAlt
    lat = -90° + rand() * 180°
    lon = -360° + rand() * 2 * 360°
    alt = rand() * 1m
    T(lat, lon, alt)
end
randpoints(::Type{T}, n) where T = map(i -> randpoint(T), 1:n)

function bench(types)
    for T in types
        println()
        println("#=============== Test $T Begin ==============#")
        println()
        println("Creating 10000 random elements of type $T (generic type, with checks):")
        b = @benchmark $randpoints($T, 10000)
        println()
        display(b)
        println()
        # Test with fully parametrized constructor, skipping checks
        P = T <: LatLonAlt ? T{WGS84Latest, Deg{Float64}, Met{Float64}} : T{WGS84Latest, Deg{Float64}}
        name = T <: LatLonAlt ? "$(T){WGS84Latest, Deg{Float64}, Met{Float64}}" : "$(T){WGS84Latest, Deg{Float64}}"
        println("Creating 10000 random elements of type $(name) (fully parametrized type, no checks):")
        b = @benchmark $randpoints($P, 10000)
        println()
        display(b)
    end
end
```

As it can be seen from the benchmark below, the checks do indeed increase construction time by ~3 times (~4 times in 1.9) in this PR compared to the unchecked inner constructor. It however seems that this implementation with `checklat` and modified `fixlon` is still faster (more so in 1.11) than the current implementation in main. I would still leave both checks for the generic constructor and give this way of skipping checks by using the fully parametrized constructor if ones want to get peak speed and can either ensure inputs are within bounds or does not care about that (probably this should be documented).

## After this PR
### Julia 1.11-rc2 - Windows
```julia-repl
julia> bench([LatLon, LatLonAlt, GeocentricLatLon, AuthalicLatLon])

#=============== Test GeodeticLatLon Begin ==============#

Creating 10000 random elements of type GeodeticLatLon (generic type, with checks):

BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  158.700 μs …   5.989 ms  ┊ GC (min … max): 0.00% … 96.35%
 Time  (median):     162.800 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   179.979 μs ± 118.746 μs  ┊ GC (mean ± σ):  4.29% ±  6.79%

  ▄█▇▄▅▁▂▁▁▁     ▂▄▃▃▂▁ ▁                                       ▁
  █████████████▆▇███████████▇▇▇▇▆▇▇▆▇▇▆▅▆▆▆▆▄▅▆▆▆▇▆█▇▄▄▅▅▄▄▄▄▃▄ █
  159 μs        Histogram: log(frequency) by time        272 μs <

 Memory estimate: 156.31 KiB, allocs estimate: 3.

Creating 10000 random elements of type GeodeticLatLon{WGS84Latest, Deg{Float64}} (fully parametrized type, no checks):

BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  54.900 μs …   7.974 ms  ┊ GC (min … max): 0.00% … 98.52%
 Time  (median):     56.800 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   71.802 μs ± 135.381 μs  ┊ GC (mean ± σ):  8.96% ±  6.22%

  █▆▅▄▂▁              ▄▄▃▃▂▂▁ ▁                                ▁
  ███████▇▇▅▇▆▆▅▆▆▅▅▆▆███████████▇▇▆▇▇▇▆▇▆▆▇▇▆▇▆▅▅▆▅▅▂▃▃▂▂▂▃▂▃ █
  54.9 μs       Histogram: log(frequency) by time       133 μs <

 Memory estimate: 156.31 KiB, allocs estimate: 3.

#=============== Test GeodeticLatLonAlt Begin ==============#

Creating 10000 random elements of type GeodeticLatLonAlt (generic type, with checks):

BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  166.600 μs …  10.119 ms  ┊ GC (min … max): 0.00% … 97.47%
 Time  (median):     171.100 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   192.962 μs ± 148.664 μs  ┊ GC (mean ± σ):  5.40% ±  7.99%

  █▆▄▂▁▁  ▅▃▂▁▁▁                                                ▁
  █████████████████▇███▆▅▅▄▄▅▅▄▄▅▄▄▄▅▄▄▃▄▄▁▃▁▄▁▁▁▃▁▁▁▁▃▄▁▃▁▁▁▁▃ █
  167 μs        Histogram: log(frequency) by time        489 μs <

 Memory estimate: 234.44 KiB, allocs estimate: 3.

Creating 10000 random elements of type GeodeticLatLonAlt{WGS84Latest, Deg{Float64}, Met{Float64}} (fully parametrized type, no checks):

BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  55.000 μs …   7.404 ms  ┊ GC (min … max):  0.00% … 98.35%
 Time  (median):     57.100 μs               ┊ GC (median):     0.00%
 Time  (mean ± σ):   74.487 μs ± 132.485 μs  ┊ GC (mean ± σ):  10.83% ±  7.49%

  █▆▄▂▂▁ ▁        ▄▄▃▁                                         ▁
  ███████████▆▆▆▅▆█████▇▇▇▇▆▆▆▅▆▅▆▅▅▅▅▅▄▁▄▄▃▅▄▅▄▄▃▅▄▁▄▃▃▄▄▄▄▄▇ █
  55 μs         Histogram: log(frequency) by time       201 μs <

 Memory estimate: 234.44 KiB, allocs estimate: 3.

#=============== Test GeocentricLatLon Begin ==============#

Creating 10000 random elements of type GeocentricLatLon (generic type, with checks):

BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  158.200 μs …   9.036 ms  ┊ GC (min … max): 0.00% … 97.39%
 Time  (median):     162.200 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   180.138 μs ± 135.100 μs  ┊ GC (mean ± σ):  4.41% ±  6.80%

  ▅█▅▅▃▃▂▂▁▁   ▅▄▃▃▂ ▁                                          ▂
  ██████████████████████▇▇▇▇▇▇▆▇█▇▆▇▇▆▅▆▅▆▆▆▆▅▅▇▅▅▅▅▅▅▅▄▃▄▁▄▄▃▃ █
  158 μs        Histogram: log(frequency) by time        293 μs <

 Memory estimate: 156.31 KiB, allocs estimate: 3.

Creating 10000 random elements of type GeocentricLatLon{WGS84Latest, Deg{Float64}} (fully parametrized type, no checks):

BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  54.800 μs …   7.896 ms  ┊ GC (min … max): 0.00% … 98.60%
 Time  (median):     56.400 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   70.358 μs ± 123.537 μs  ┊ GC (mean ± σ):  8.59% ±  6.23%

  █▆▃▃▁            ▄▄▃▂▁                                       ▁
  █████████▇██▇▆▆▅▇█████▇▇▆▇▇▅▇▆▆▆▅▅▅▅▄▅▅▆▄▄▅▄▅▃▄▄▄▃▄▃▃▄▃▃▂▃▄▄ █
  54.8 μs       Histogram: log(frequency) by time       148 μs <

 Memory estimate: 156.31 KiB, allocs estimate: 3.

#=============== Test AuthalicLatLon Begin ==============#

Creating 10000 random elements of type AuthalicLatLon (generic type, with checks):

BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  158.300 μs …   5.586 ms  ┊ GC (min … max): 0.00% … 96.06%
 Time  (median):     162.800 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   178.875 μs ± 103.934 μs  ┊ GC (mean ± σ):  3.95% ±  6.77%

  ▄█▅▅▃▂▂▂▁▁  ▁▄▂▂▁ ▁                                           ▁
  ████████████████████▇▇▇▇▇▆▇▇█▇▇▆▇▆▆▅▅▄▅▅▅▅▃▃▃▄▄▃▂▄▃▄▄▅▄▄▂▃▄▃▅ █
  158 μs        Histogram: log(frequency) by time        301 μs <

 Memory estimate: 156.31 KiB, allocs estimate: 3.

Creating 10000 random elements of type AuthalicLatLon{WGS84Latest, Deg{Float64}} (fully parametrized type, no checks):

BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  54.900 μs …  13.777 ms  ┊ GC (min … max): 0.00% … 99.17%
 Time  (median):     56.400 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   70.921 μs ± 164.284 μs  ┊ GC (mean ± σ):  9.16% ±  6.22%

  █▆▄▃▂▁             ▁▃▄▃▂▁▁                                   ▁
  ███████▇▇▇▇▆▄▇▆▇▅▆▆████████▇▇████▇▇▆▆▆▆▆▆▆▅▅▅▄▅▅▅▅▄▄▄▅▅▄▄▃▄▅ █
  54.9 μs       Histogram: log(frequency) by time       134 μs <

 Memory estimate: 156.31 KiB, allocs estimate: 3.
```

### Julia 1.9.4 - Windows
```julia-repl
julia> bench([LatLon, LatLonAlt, GeocentricLatLon, AuthalicLatLon])

#=============== Test GeodeticLatLon Begin ==============#

Creating 10000 random elements of type GeodeticLatLon (generic type, with checks):

BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  213.500 μs …   3.918 ms  ┊ GC (min … max): 0.00% … 91.68%
 Time  (median):     226.500 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   243.076 μs ± 110.634 μs  ┊ GC (mean ± σ):  2.20% ±  4.71%

  ▇█▅▆▄▄▄▂▁▁▁▄▇▅▅▄▃▃▃▁▂▂▁▁▁▁  ▁                                 ▂
  ████████████████████████████████▇▇█▇▇█▇████▇▆▆▅▆▆▆▅▆▇▆▅▇▇▅▆▃▆ █
  214 μs        Histogram: log(frequency) by time        360 μs <

 Memory estimate: 156.30 KiB, allocs estimate: 2.

Creating 10000 random elements of type GeodeticLatLon{WGS84Latest, Deg{Float64}} (fully parametrized type, no checks):

BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  55.500 μs …  3.625 ms  ┊ GC (min … max): 0.00% … 96.65%
 Time  (median):     58.000 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   73.320 μs ± 98.137 μs  ┊ GC (mean ± σ):  6.46% ±  5.14%

  █▇▅▂▂▂▁▁         ▄▆▅▃▂▁▁▁                                   ▂
  ████████▇██▇▇▅▇▇▄█████████▇▇▇▇▆▆▇▆▆▆▅▆▅▆▆▆▅▅▅▅▅▆▅▆▄▅▅▄▄▃▄▄▅ █
  55.5 μs      Histogram: log(frequency) by time       146 μs <

 Memory estimate: 156.30 KiB, allocs estimate: 2.

#=============== Test GeodeticLatLonAlt Begin ==============#

Creating 10000 random elements of type GeodeticLatLonAlt (generic type, with checks):

BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  215.700 μs …   5.852 ms  ┊ GC (min … max): 0.00% … 95.38%
 Time  (median):     224.300 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   248.930 μs ± 136.460 μs  ┊ GC (mean ± σ):  2.89% ±  5.53%

  ██▇▅▄▃▂▂▁▁▁ ▂▆▄▅▃▂▂▂▂▁▁▁▁                                     ▂
  █████████████████████████████▇▇▇▇▇▇▆▇▇▆▇▇██▇▇▅▅▆▆▄▅▄▄▄▄▅▄▄▄▅▅ █
  216 μs        Histogram: log(frequency) by time        414 μs <

 Memory estimate: 234.42 KiB, allocs estimate: 2.

Creating 10000 random elements of type GeodeticLatLonAlt{WGS84Latest, Deg{Float64}, Met{Float64}} (fully parametrized type, no checks):

BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  66.700 μs …   6.737 ms  ┊ GC (min … max): 0.00% … 98.05%
 Time  (median):     68.700 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   87.168 μs ± 114.553 μs  ┊ GC (mean ± σ):  6.80% ±  6.04%

  █▆▃▂▁▁      ▁▁    ▂▄▄▂                                       ▁
  █████████▇▆████▅▆▆█████▇▇▇▇▇▇▅▆▅▅▅▆▆▆▅▆▆▆▅▅▄▅▅▅▄▃▄▅▄▅▅▄▄▄▄▅▃ █
  66.7 μs       Histogram: log(frequency) by time       199 μs <

 Memory estimate: 234.42 KiB, allocs estimate: 2.

#=============== Test GeocentricLatLon Begin ==============#

Creating 10000 random elements of type GeocentricLatLon (generic type, with checks):

BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  212.600 μs …   5.737 ms  ┊ GC (min … max): 0.00% … 95.31%
 Time  (median):     219.000 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   234.399 μs ± 117.583 μs  ┊ GC (mean ± σ):  2.02% ±  4.43%

  ▃█▇▆▅▄▃▃▂▂▁▁▁▄▃▂▂▁▁▁  ▁                                       ▂
  ████████████████████████████▇▇▇▇▇▇▆▆▄▆▆▅▅▆▇▅▅▅▆▅▆▆▅▅▅▆▇▅▅▃▅▄▄ █
  213 μs        Histogram: log(frequency) by time        352 μs <

 Memory estimate: 156.30 KiB, allocs estimate: 2.

Creating 10000 random elements of type GeocentricLatLon{WGS84Latest, Deg{Float64}} (fully parametrized type, no checks):

BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  55.300 μs …  3.585 ms  ┊ GC (min … max): 0.00% … 96.94%
 Time  (median):     57.400 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   66.606 μs ± 77.183 μs  ┊ GC (mean ± σ):  5.21% ±  4.97%

  ▆█▆▅▃▂▁▁▁▁                ▂▁▃▂▁                             ▁
  ████████████▇▆█▇▆▅▆▅▅▅▅▇▇█████████▇▇▇▇▇▆▇▇▆▆▆▅▅▅▆▅▅▄▄▅▄▄▅▄▅ █
  55.3 μs      Histogram: log(frequency) by time       119 μs <

 Memory estimate: 156.30 KiB, allocs estimate: 2.

#=============== Test AuthalicLatLon Begin ==============#

Creating 10000 random elements of type AuthalicLatLon (generic type, with checks):

BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  213.600 μs …   7.194 ms  ┊ GC (min … max): 0.00% … 96.19%
 Time  (median):     218.400 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   232.503 μs ± 126.267 μs  ┊ GC (mean ± σ):  2.01% ±  4.38%

  ▄█▇▅▅▃▃▃▂▁▂▁ ▂▃▂▂▂▁ ▁                                         ▂
  ██████████████████████▇█▇▇▇██▆█▇▇▇▆▇▇▆▆▇▅▇▆▆▆▇▆▆▅▆▆▅▇█▆▄▃▅▅▁▄ █
  214 μs        Histogram: log(frequency) by time        346 μs <

 Memory estimate: 156.30 KiB, allocs estimate: 2.

Creating 10000 random elements of type AuthalicLatLon{WGS84Latest, Deg{Float64}} (fully parametrized type, no checks):

BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  55.200 μs …  3.361 ms  ┊ GC (min … max): 0.00% … 95.16%
 Time  (median):     57.200 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   66.459 μs ± 74.773 μs  ┊ GC (mean ± σ):  4.94% ±  4.95%

  ▇█▅▃▂▁           ▂▃▃▂▁                                      ▁
  ████████▇██▇▇▇▇▅▆██████▇▇▆▆▆▆▅▆▆▆▅▅▅▄▅▆▄▅▅▃▂▅▅▅▅▄▃▅▃▄▄▃▄▄▄▆ █
  55.2 μs      Histogram: log(frequency) by time       148 μs <

 Memory estimate: 156.30 KiB, allocs estimate: 2.
```

## Before PR (main)
### Julia 1.11-rc2 - Windows
```julia-repl
julia> bench([LatLon, LatLonAlt, GeocentricLatLon, AuthalicLatLon])

#=============== Test GeodeticLatLon Begin ==============#

Creating 10000 random elements of type GeodeticLatLon (generic type, with checks):

BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  248.400 μs …   8.571 ms  ┊ GC (min … max): 0.00% … 96.19%
 Time  (median):     256.500 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   275.683 μs ± 153.088 μs  ┊ GC (mean ± σ):  3.50% ±  6.79%

  █▃▆▃▃▃▁▆▅▄▃▃▂▂▁▁                                              ▂
  █████████████████▇██▇▇▆▇▇▆▅▆▅▅▅▄▅▄▁▄▄▄▄▃▅▁▄▅▄▃▃▅▅▃▄▁▁▄▁▃▁▃▁▁▃ █
  248 μs        Histogram: log(frequency) by time        478 μs <

 Memory estimate: 156.31 KiB, allocs estimate: 3.

Creating 10000 random elements of type GeodeticLatLon{WGS84Latest, Deg{Float64}} (fully parametrized type, no checks):

BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  55.000 μs …  11.977 ms  ┊ GC (min … max): 0.00% … 99.12%
 Time  (median):     57.200 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   74.165 μs ± 162.947 μs  ┊ GC (mean ± σ):  9.30% ±  6.23%

  █▆▄▂▂▁▁   ▁     ▅▄▄▂▁    ▁▁▁                                 ▁
  ████████▇██▇▇▆▆████████▇▇████▇▇▅▆▅▄▅▄▅▅▃▄▃▅▃▄▄▃▅▃▅▄▄▅▄▂▄▄▅▇▆ █
  55 μs         Histogram: log(frequency) by time       153 μs <

 Memory estimate: 156.31 KiB, allocs estimate: 3.

#=============== Test GeodeticLatLonAlt Begin ==============#

Creating 10000 random elements of type GeodeticLatLonAlt (generic type, with checks):

BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  251.800 μs …   6.544 ms  ┊ GC (min … max): 0.00% … 94.92%
 Time  (median):     254.000 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   277.546 μs ± 138.184 μs  ┊ GC (mean ± σ):  4.18% ±  7.95%

  █▄▃▅▂▁▁                                                       ▁
  ████████▇▇▇▅▄▁▄▃▁▅▁▃▁▁▁▁▁▁▁▁▁▃▁▁▁▁▁▁▁▃▁▁▁▁▃▁▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▄▄ █
  252 μs        Histogram: log(frequency) by time        1.1 ms <

 Memory estimate: 234.44 KiB, allocs estimate: 3.

Creating 10000 random elements of type GeodeticLatLonAlt{WGS84Latest, Deg{Float64}, Met{Float64}} (fully parametrized type, no checks):

BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  55.100 μs …   7.335 ms  ┊ GC (min … max):  0.00% … 98.31%
 Time  (median):     56.200 μs               ┊ GC (median):     0.00%
 Time  (mean ± σ):   74.150 μs ± 135.312 μs  ┊ GC (mean ± σ):  10.69% ±  7.44%

  █▅▂▁▁▁▁           ▃▃▃▁▁                                      ▁
  ████████▆▆▄▅▆▆▆▄▆▅█████▇▇▇▇▇▆▅▅▆▆▆▅▆▅▄▅▅▄▃▅▅▄▃▂▄▃▄▃▃▃▄▅▃▃▂▄▃ █
  55.1 μs       Histogram: log(frequency) by time       185 μs <

 Memory estimate: 234.44 KiB, allocs estimate: 3.

#=============== Test GeocentricLatLon Begin ==============#

Creating 10000 random elements of type GeocentricLatLon (generic type, with checks):

BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  250.400 μs …   4.696 ms  ┊ GC (min … max): 0.00% … 93.40%
 Time  (median):     257.200 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   271.079 μs ± 109.544 μs  ┊ GC (mean ± σ):  3.04% ±  6.76%

  █▄▇▄▃▃▂▂▂▅▂▁▁▁                                                ▂
  ██████████████████▆▇▆▇▆▆▆▆▅▄▅▅▄▅▃▅▁▅▅▁▄▁▆▄▅▅▅▄▁▅▁▄▃▃▃▃▃▄▃▁▃▁▃ █
  250 μs        Histogram: log(frequency) by time        447 μs <

 Memory estimate: 156.31 KiB, allocs estimate: 3.

Creating 10000 random elements of type GeocentricLatLon{WGS84Latest, Deg{Float64}} (fully parametrized type, no checks):

BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  55.000 μs …  10.956 ms  ┊ GC (min … max): 0.00% … 98.94%
 Time  (median):     56.600 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   69.951 μs ± 142.007 μs  ┊ GC (mean ± σ):  8.85% ±  6.11%

  █▆▆▄▃▂▁               ▁▄▃▃▃▁                                 ▁
  █████████▆▅▅▆▅▅▅▅▄▃▄▃▄███████▇▇▇▇█▇▇▇▇▆▅▅▆▅▆▆▅▅▆▅▅▅▅▅▅▅▅▅▄▄▄ █
  55 μs         Histogram: log(frequency) by time       124 μs <

 Memory estimate: 156.31 KiB, allocs estimate: 3.

#=============== Test AuthalicLatLon Begin ==============#

Creating 10000 random elements of type AuthalicLatLon (generic type, with checks):

BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  250.200 μs …   4.551 ms  ┊ GC (min … max): 0.00% … 93.22%
 Time  (median):     253.000 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   269.488 μs ± 117.688 μs  ┊ GC (mean ± σ):  3.26% ±  6.84%

  ▇█▃▃▄▂▃▂▂▁▁  ▄▄▂▁▂▁                                           ▂
  ████████████▇██████████▇█▇▇▇▆▇▆▅▆▅▅▅▆▅▅▅▅▅▄▄▅▄▅▅▅▄▃▃▅▁▁▄▄▃▃▄▃ █
  250 μs        Histogram: log(frequency) by time        379 μs <

 Memory estimate: 156.31 KiB, allocs estimate: 3.

Creating 10000 random elements of type AuthalicLatLon{WGS84Latest, Deg{Float64}} (fully parametrized type, no checks):

BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  54.900 μs …   9.739 ms  ┊ GC (min … max): 0.00% … 98.42%
…