@disberd 2024-08-03
@juliohm, considering the significant performance boost I see from PR #147, I think it would be best to redo benchmarks with `rand` fixed as current benchmarks might be skewed as the time for generating random LatLon in main has significant overhead due to allocations/type instability.