@disberd 2024-08-04
I modified the benchmarks to avoid getting biased by `rand` allocations and revert back to do lat/lon checks only in the more specialized outer constructor to give an opt-out for checks if performance is critical.

New benchmarks (I skipped 1.10 for sake of time/readability) show that this new implementation despite checking for lat and also avoiding the need of `islon180` can still be slightly faster than the previous version of `fixlon` only