@juliohm 2024-08-05
@disberd we introduced two new functions to make sure that algorithms always return valid lat/lon pairs and rebased this PR to take advantage of this. We can now remove the arbitrary `1e-6` tolerance in the `checklat`.