@juliohm 2024-08-05
Ran the benchmark locally against the main branch and performance is the same on my machine with Julia v1.10