@eliascarv 2024-08-05
@disberd, can you please add some tests for `fixlon`?
In the "GeodeticLatLon" testset, above the "show" test, you can add some tests like this:
```julia
# fix longitude
@test LatLon(T(0), T(225)) == LatLon(T(0), T(-135))
... # some more tests

c = LatLon(T(1), T(2))
@test sprint(show, c) == "GeodeticLatLon{WGS84Latest}(lat: 1.0°, lon: 2.0°)"
```